pas mal de corrections

+++ extracteur.sh
@@ -17,33 +17,38 @@ ARGUMENTS=( "" "" x -xvf -zxvf -zxvf -jxvf e )
 mkdir NIMEWF
 for((i=1;i<${#EXTENSION[*]};i++))
   do
-    for FILE in `ls *.${EXTENSION[$i]} 2>> $HOME/logs/nimscripts.log | sed "s/.${EXTENSION[$i]}//"`
-      do
-	mkdir $FILE
-	mv $FILE.${EXTENSION[$i]} $FILE/
-	cd $FILE/
-	${PROGRAMME[$i]} ${ARGUMENTS[$i]} "$FILE.${EXTENSION[$i]}"
-	NB=2
-	if [[ "$1" == *r* ]]
-	  then
-	    rm $FILE.${EXTENSION[$i]}
-	    NB=1
-	  fi
-	if [[ `ls | wc -l` == $NB ]]
-	  then
-	    mv * ../NIMEWF
-	    cd ..
-	    rmdir $FILE/
-	  else
-	    cd ..
-	  fi
-    done
+    if [ -e *.${EXTENSION[$i]} ]
+      then
+        for FILE in `ls *.${EXTENSION[$i]} | sed "s/.${EXTENSION[$i]}//"`
+          do
+	    mkdir $FILE
+	    mv $FILE.${EXTENSION[$i]} $FILE/
+	    cd $FILE/
+	    ${PROGRAMME[$i]} ${ARGUMENTS[$i]} "$FILE.${EXTENSION[$i]}"
+	    NB=2
+	    if [[ "$1" == *r* ]]
+	      then
+	        rm $FILE.${EXTENSION[$i]}
+	        NB=1
+	      fi
+	    if [[ `ls | wc -l` == $NB ]]
+	      then
+	        mv * ../NIMEWF
+	        cd ..
+	        rmdir $FILE/
+	      else
+	        cd ..
+	      fi
+        done
+      fi
 done
 
-mv NIMEWF/* . 2>> $HOME/logs/nimscripts.log
+if [ -e NIMEWF/* ]
+  then
+    mv NIMEWF/* .
+  fi
 rmdir NIMEWF
 
-$HOME/scripts/nimscriptslog.sh
 IFS=$OLDIFS
 
 exit 0
